fix: satisfy mypy no-any-return in the backoff delay helper
Follow-up to the keepalive commit's _backoff_delay extraction; also
trims a line-length violation in its regression test.

+++ custom_components/hyperhdr/client.py
@@ -74,7 +74,7 @@ _MAX_BACKOFF_ATTEMPT = 20
 
 def _backoff_delay(attempt: int) -> float:
     """The deterministic (pre-jitter) reconnect backoff delay for `attempt`."""
-    return min(RECONNECT_BASE_DELAY * (2 ** min(attempt, _MAX_BACKOFF_ATTEMPT)), RECONNECT_MAX_DELAY)
+    return float(min(RECONNECT_BASE_DELAY * (2 ** min(attempt, _MAX_BACKOFF_ATTEMPT)), RECONNECT_MAX_DELAY))
 
 
 class HyperHdrBaseClient:
